requirements.txt
# Function dependencies, for example:
# package>=version
facebook-sdk
pandas
matplotlib
google.cloud